requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
zope.event==4.5.0
zope.interface==5.4.0
